refactor: extract ui state module
