+++ src/core/message.rs
@@ -1,4 +1,4 @@
-#[derive(Clone)]
+#[derive(Debug, Clone)]
 pub struct Message {
     pub role: String,
     pub content: String,
